backend/requirements.txt
flask
flask-cors
pandas
openai==0.28
scikit-learn
python-dotenv